Check user input for observations and measurement
Checks have been added for the measurement and observations arguments.
They will raise an error if measurement is not a valid measurement.
They will raise a warning if some column in observations is not defined
terminology.

Correctly handles prefixed terms.

The corresponding validators are in
analysis_neuro._validate_observations and
analysis_neuro._validate_measurement