Remove @measures decorator.
For now, builtin measurements will just have to remain functions that modelers can use inside their models. It would be nice for the framework to be able to automatically pick the appropriate method for a given measurement on a given model, without the user necessarily needing to search through the available methods, however as I cannot think of a satisfactory way to do this and it is not immediately relevant, we will leave this as a later refinement.