Draft cossell 2015 psp validation
One uncertain thing is how exactly to deal with paired measurements.

We could expect the measurement methods to provide a relevant index.
However, in this example with response correlation one of the measurements
(correlation) does not have pre/post while the other does.
So this makes joining them based on index not work.
It may be that we need pair-specific joinings.
Maybe a default that works for most pairs by relying on index?

For now we have implemented this by just expecting the model to provide
response correlations and psps in equal amounts in separate measurements.
This will do for purpose-made quick and dirty model classes, but it will
not scale to a more general model.